+++ src/komm/__init__.py
@@ -75,6 +75,7 @@ from ._finite_state_machine.MooreMachine import MooreMachine
 from ._integer_coding.EliasDeltaCode import EliasDeltaCode
 from ._integer_coding.EliasGammaCode import EliasGammaCode
 from ._integer_coding.FibonacciCode import FibonacciCode
+from ._integer_coding.TruncatedBinaryCode import TruncatedBinaryCode
 from ._integer_coding.UnaryCode import UnaryCode
 
 # Labelings
@@ -200,6 +201,7 @@ __all__ = [
     "EliasDeltaCode",
     "EliasGammaCode",
     "FibonacciCode",
+    "TruncatedBinaryCode",
     "UnaryCode",
     # _labelings
     "Labeling",